+++ testWrapper/tests/RmsSensor_test.cpp
@@ -6,7 +6,7 @@
 NTFX_TEST_BEGIN
 
 NTFX_TEST() {
-  auto rmsSensor = NtFx::RmsSensorStereo<float>();
+  auto rmsSensor = NtFx::RmsSensorStereo<double>();
   rmsSensor.setT_ms(10);
   NTFX_ADD_TEST(rmsSensor, "dynamic");
   return NTFX_RUN_TESTS();
