+++ testWrapper/tests/SoftClip_test.cpp
@@ -6,9 +6,9 @@
 NTFX_TEST_BEGIN
 
 NTFX_TEST() {
-  auto softClip3 = NtFx::SoftClip3<float>();
+  auto softClip3 = NtFx::SoftClip3<double>();
   NTFX_ADD_TEST(softClip3, "linearSweep");
-  auto softClip5 = NtFx::SoftClip5<float>();
+  auto softClip5 = NtFx::SoftClip5<double>();
   NTFX_ADD_TEST(softClip5, "linearSweep");
   return NTFX_RUN_TESTS();
 }
